fix: reusable search hook and bug fix
Reusable search hook, constant separation and bug fix.
