+++ src/app/[locale]/(unauth)/page.tsx
@@ -1,75 +1,28 @@
 'use client';
 
-import { useQuery } from '@apollo/client';
 import { Box, SimpleGrid, Text } from '@chakra-ui/react';
 import { useTranslations } from 'next-intl';
-import React, { useCallback, useEffect, useState } from 'react';
-import { useDebounce } from 'use-debounce';
+import React, { useCallback, useEffect } from 'react';
 
 import PodcastCard from '@/components/Podcast';
 import PodcastSkeleton from '@/components/PodcastSkeleton';
-import { GET_CONTENT_CARDS } from '@/libs/apollo/getContent';
+import useContentCards from '@/hooks/useContentCards';
 import { useSearch } from '@/libs/context/SearchProvider';
 
 // Define types for the content cards
 const Index = () => {
   const { searchTerm } = useSearch();
-  const [debouncedSearchTerm] = useDebounce(searchTerm, 300);
-  const [offset, setOffset] = useState(0); // State for pagination offset
-  const [contentCards, setContentCards] = useState<Podcast[]>([]); // State to store fetched data
   const t = useTranslations('MainPage');
 
-  // GraphQL query with search term and pagination variables
-  const { loading, refetch, data } = useQuery<{
-    contentCards: ContentCard;
-  }>(GET_CONTENT_CARDS, {
-    variables: { keyword: debouncedSearchTerm, limit: 10, offset },
-    fetchPolicy: 'cache-and-network', // Ensure we always have fresh data
-  });
+  const { loading, contentCards, setOffset, offset, data } =
+    useContentCards(searchTerm);
 
-  const fetchContentCards = useCallback(async () => {
-    try {
-      const { data: refetchData } = await refetch({
-        keyword: debouncedSearchTerm,
-        limit: 10,
-        offset,
-      });
-      if (offset === 0) {
-        setContentCards(refetchData.contentCards.edges);
-      } else {
-        setContentCards((prevCards) => [
-          ...prevCards,
-          ...refetchData.contentCards.edges,
-        ]);
-      }
-    } catch (error) {
-      console.error('Error fetching content cards:', error);
-    }
-  }, [debouncedSearchTerm, offset, refetch]);
-
-  useEffect(() => {
-    if (data?.contentCards?.edges && offset === 0) {
-      setContentCards((prevCards) => [
-        ...prevCards,
-        ...data.contentCards.edges,
-      ]);
-    }
-  }, [data]);
-
-  // Fetch content cards initially and on search term or offset change
-  useEffect(() => {
-    if (debouncedSearchTerm) setOffset(0);
-    fetchContentCards();
-  }, [debouncedSearchTerm, fetchContentCards]);
-
-  // Function to handle infinite scrolling
   const handleScroll = useCallback(async () => {
     if (
       window.innerHeight + document.documentElement.scrollTop !==
       document.documentElement.offsetHeight - 0.5
     )
       return;
-    // setOffset((prevOffset) => prevOffset + 10);
     if (contentCards.length < (data?.contentCards.meta.total || 0)) {
       setOffset((prevOffset) => prevOffset + 10);
     }
