+++ src/utils/Constants.ts
@@ -1,0 +1,4 @@
+const SEARCH_TIMEOUT = 300;
+const PAGE_LIMIT = 10;
+
+export { PAGE_LIMIT, SEARCH_TIMEOUT };
